+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## 0.2.0 (TBD)
+
+- Add a new `pc` (Pydantic context) command, displaying useful
+  information about the current schema generation context.
+- Simplify handling of the Rich console.
+- Use [`Mapping`](https://docs.python.org/3/library/collections.abc.html#collections.abc.Mapping)
+  and [`Sequence`](https://docs.python.org/3/library/collections.abc.html#collections.abc.Sequence)
+  from the [`collections.abc`](https://docs.python.org/3/library/collections.abc.html) module
+  when cleaning the core schema.
+
 ## 0.1.0 (2024-09-03)
 
 Initial release.